backend/requirements.txt
fastapi==0.110.0
uvicorn==0.27.1
python-dotenv==1.0.1
spotipy==2.23.0
weaviate-client==4.11.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.9
pydantic>=2.8.0,<3.0.0
httpx==0.27.0 